+++ gpx_linesman/__init__.py
@@ -1,2 +1,1 @@
 from .cli import run
-