+++ gpx_linesman/measure.py
@@ -5,8 +5,9 @@ from .util import project
 
 
 
-class ScoringMethod:
+class Measure:
     """Abstract base class evaluating a sequence of points to a straight line."""
+    desc = None  # description of the value aggregated by this measure
 
     def __init__(self, points, line_start, line_end):
         """
@@ -34,7 +35,7 @@ class ScoringMethod:
         raise NotImplementedError()
 
 
-class MeterDeviation(ScoringMethod):
+class MeterDeviation(Measure):
     """
     Scoring method using the deviation of points from the line in meters as
     critical measure.
@@ -47,6 +48,8 @@ class MeterDeviation(ScoringMethod):
 
 class MaxDeviation(MeterDeviation):
     """Maximum deviation from the line in meters."""
+    desc = 'Maximum deviation in meters'
+
     def aggregate(self):
         return max(self.measures)
 
@@ -54,15 +57,18 @@ class MaxDeviation(MeterDeviation):
 
 class AvgDeviation(MeterDeviation):
     """Average deviation from the line in meters."""
+    desc = 'Average deviation in meters'
+
     def aggregate(self):
         return sum(self.measures)/len(self.measures)
 
 
-class AvgSquareDeviation(MeterDeviation):
+class SquareDeviationAvg(MeterDeviation):
     """
     Average squared deviation from the line in meters^2.
     This measure punishes bigger deviations more than small ones.
     """
+    desc = 'Average squared deviation'
+
     def aggregate(self):
         return sum(map(lambda x: x**2, self.measures))/len(self.measures)
-