+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  - Replace `linestart` and `lineend` arguments defining the straight line with a
    `--line`/`-l` argument. If omitted, the first and last point in the gpx file
    are used to define the reference line.
+ - Replace `--using` argument with a positional `measure` argument
+ - Change identifiers for measures
 
 ## 0.1.2